refactor: Extract TopicPrefix to a module under models
